free_ebook_maker/requirements.txt
ddgs
requests
google-generativeai
pypandoc
